examples/requirements.txt
flask
pycups